pyproject.toml
[project]
name = "feature-transform"
version = "1.0.0"
description = "Build Scikit ColumnTransformers by specifying configs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "polars>=1.0.0",
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.6",
]

[tool.pytest.ini_options]
addopts = [
    '--verbose',
    "--cov=feature_transform",
    "--cov-report=html",
    "--cov-report=term",
]
filterwarnings = "ignore"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["feature_transform"]
